Fix error details from VBScript in some cases
Fix a case where error details including the stack trace were not getting reported when they should. Also added a test case that proves that the error details in this case get reported correctly.

Note, error handling is quite complex in here, so I'm not 100% certain this is always the right thing to do. We also have some indication that the problem may be intermittent, though I haven't been able to prove that yet.