Fix fast up to date check (#211)
* Remove non-existent folders

These folders don't actually exist on disk. They are added automatically to the tree for linked items, but listing them explicitly in the project when they don't exist on disk causes them to be shown with error overlays in VS 16.9.

* Use consistent target frameworks

This works around an issue that can cause VS to needlessly rebuild these two projects.