feat(WindowsScriptEngine): SynchronizationContext variant
`WindowsScriptEngine` uses `System.Windows.Threading.Dispatcher`. On .NET Core or .NET 5, this limits its usage to WindowsDesktop projects that use WPF.

Here we add a variant that uses the current `SynchronizationContext` instead. This allows to use the engine in all types of projects.

In WindowsDesktop projects that use WPF, the default `SynchronizationContext` is a `DispatcherSynchronizationContext`, so this will keep the same behaviour as the Dispatcher-based version: it maintains thread-affinity by using the Windows message queue for the current thread.
In WindowsDesktop projects that use WinForms, the default `SynchronizationContext` is a `WindowsFormsSynchronizationContext` with the same behaviour.

In other projects, the requirements guide the choice of a `SynchronizationContext`:
- if thread-affinity is not required, callers can use the default `SynchronizationContext` which is based on the threadpool. When there is no current `SynchronizationContext`, `WindowsScriptEngine` will create an instance of the default `SynchronizationContext`.
- if thread-affinity is required, callers can setup a single-thread `SynchronizationContext`,
- if thread-affinity is required and the Windows message queue needs to be processed (for example to use COM objects like the async `MSXML2.XMLHTTP` `XMLHttpRequest`), callers can setup a `SynchronizationContext` based on the Windows message queue, such as the `MessageQueueSynchronizationContext` implemented in the unit tests.

To avoid changing the existing ClearScript distribution, the `WindowsScriptEngine` variant is implemented in a new `ClearScript.WindowsSyncContext` project. A new test project `ClearScriptTest.SyncContext` is also created, it is a copy of `ClearScriptTest` that uses the `WindowsScriptEngine` variant based on the `SynchronizationContext`.

In this project, we have implemented 2 versions of the `***ScriptEngine_AddCOMType_XMLHTTP` unit test:
- one which uses the WPF `DispatcherSynchronizationContext`
- one which uses the `MessageQueueSynchronizationContext` implemented in the unit test project

These new projects target .NET Core and .NET 5 (`ClearScript.WindowsSyncContext` could also target netstandard2.0).

Test plan: all unit test pass