WindowsScriptEngine: abstract the Dispatcher, allow usage without the WindowsDesktop runtime
`WindowsScriptEngine` uses `System.Windows.Threading.Dispatcher`. On .NET Core or .NET 5, this requires the WindowsDesktop runtime to be installed on the host.

Here we abstract the use of the `Dispatcher` in an `ISyncInvoker` interface, so that `ClearScript.Windows` can be distributed without the dependency on the WindowsDesktop runtime. `ISyncInvoker` is an interface that represents the fact that Windows Script engines require thread-affinity.

Constructor overloads are added to `JScriptEngine` and `VBScriptEngine` to pass an instance of `ISyncInvoker`.

Constructors that take no `ISyncInvoker` parameter create a `DefaultSyncInvoker`, which is a very simple implementation suitable for scripts that do not perform any form of event handling.

A separate `DispatcherSyncInvoker` is provided for the other use cases. It is distributed in a separate `ClearScript.WindowsDesktop` project. It is implemented using a `System.Windows.Threading.Dispatcher`, which is suitable for scripts that use event handling, or scripts that need the Windows message queue to be processed.

Test plan:
- all unit test pass
- `ClearScript.Windows` can be referenced in another project that does not enables `UseWPF`, and can be run on a host where the WindowsDesktop runtime is not installed